Auditd work
* Per best practices, auditd is set to switch to single user mode on
disk being full. This is a required security action in my mind, as on a
compromised system we either want full logging, or a security measure in
place in case an attacker floods storage.
* Other auditd.conf settings have been configured based on preliminary
investigation into best practices. I imagine these will need to be
adjusted some more as we continue research.
* Baseline rule set will probably need to be adjusted as well.