Add in Cloudwatch package install
* Still need to work on configuration of Cloudwatch. Going to have to
dive into the configuration file to understand what we need to set up
here. Removing critical_files.md was fine - we can figure out that
problem in the future.
* Reminder that we're putting all tagging into the main.yml for each
role, as well as conditionals for running. This makes it easier to
separate than if we put conditionals in the playbook. Adding a tag there
means that we can force the execution even if the variable is set to not
run